Fixed bug with creating two SFFs in different compute nodes
When appending a new SFF to a list, a pointer to dataplane-ip was being passed which caused
errors when the original reference was updated, which is the case for more than one SFF node.

This has been fixed by passing a (deep) copy of the SFF dictionary to the list.